attendance_system/requirements.txt
flask
psycopg2-binary
gunicorn
openpyxl
schedule
APScheduler
pytz